Allow site.posts_attr_hash to accept non-Enumerables
The previous implementation of `site.posts_attr_hash` applied the `.each` method to the indexing attribute. If that attribute was not Enumerable, the method throws an error and Jekyll fails to build. This implementation checks whether the attribute responds to `each` and if not, uses the attribute itself as a key in the generated hash.